requirements.txt
torch>=1.7.1
torchvision>=0.8.2
gdown>=3.13.0
scipy>=1.6.0
opencv-python>=4.5.1.48
Pillow>=8.2.0
tqdm>=4.61.1
pandas>=1.2.4
numpy>=1.20.3
h5py>=3.3.0
wget>=3.2